requirements.txt
# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Database
sqlalchemy>=2.0.0
sqlparse>=0.4.4

# Embeddings and Vector Operations
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API (Optional - for future)
fastapi>=0.109.0
uvicorn>=0.27.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Streamlit Web Interface
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.1.0
